feat/ authorization with file system

+++ src/main/java/diploma/cloudapi/web/dto/FileListResponse.java
@@ -5,11 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+import java.util.Date;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class FileListResponse {
     private String filename;
-    private Integer size;
+    private Long size;
 }
